+++ vue-i18n-extract.config.js
@@ -1,5 +1,5 @@
 module.exports = {
-  "vueFiles": "{udata_front,udata-front-plugins-helper}/**/{assets/js,src}/**/*.{js,vue}",
+  "vueFiles": "udata_front/**/assets/js/**/*.{js,ts,vue}",
   "languageFiles": "udata_front/theme/**/js/locales/**/*.?(json|yaml|yml|js)",
   "exclude": [],
   "output": false,