+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Add new dataset card [#445](https://github.com/datagouv/udata-front/pull/445)
 - Fix tests for oembed CORS [#453](https://github.com/datagouv/udata-front/pull/453)
 - Add reports [#436](https://github.com/datagouv/udata-front/pull/436)
+- Switch to @datagouv/components [#439](https://github.com/datagouv/udata-front/pull/439)
 
 ## 5.0.1 (2024-06-12)