+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Fix tests for oembed CORS [#453](https://github.com/datagouv/udata-front/pull/453)
 - Add reports [#436](https://github.com/datagouv/udata-front/pull/436)
 - Switch to @datagouv/components [#439](https://github.com/datagouv/udata-front/pull/439)
+- Update translations [#458](https://github.com/datagouv/udata-front/pull/458)
 
 ## 5.0.1 (2024-06-12)